refactor: split advisor model analysis
